requirements.txt
bs4
requests
language-tool-python
nltk
transformers
torch
openai
accelerate
numpy
python-Levenshtein
pyspark